requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-standardized-errors==0.14.1
importlib_metadata==8.2.0
Markdown==3.7
sqlparse==0.5.1
typing_extensions==4.12.2
zipp==3.20.0
